feat: Refactor grant
Closes: #55
